eve/requirements.txt
flake8==2.6.0
buildbot[tls]==0.9.0b9
buildbot-pkg==0.9.0b9
buildbot-console-view==0.9.0b9
buildbot-worker==0.9.0b9
buildbot-waterfall-view==0.9.0b9
buildbot-www==0.9.0b9
docker-py==1.8.1
paramiko==2.0.1
pyaml==15.8.2
isort==4.2.5
python-novaclient==2.35
requests[security]==2.10.0
service_identity==16.0.0